pyproject.toml
[project]
name = "museum-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "watchdog>=6.0.0",
]
